Modularized SubletImage Model, SubletManager, LocationManager
